merge develop to master

+++ admin/sql/install.mysql.utf8.sql
@@ -93,7 +93,7 @@ CREATE TABLE `#__lupo_game_genre` (
 CREATE TABLE `#__lupo_game_related` (
   `id` int(11) NOT NULL AUTO_INCREMENT,
   `gameid` int(11) DEFAULT NULL,
-  `number` int(11) DEFAULT NULL,
+  `number` double DEFAULT NULL,
   PRIMARY KEY (`id`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8
